Initial commit on puppetlabs_repo
Current iteration allows for client-side configuration of yum
repository for yum.puppetlabs.com.  The base class is puppetlabs_repo
and then puppetlabs_repo::yum is available to setup yum.  It is included
in a somewhat smarter way, using facts to see if the system has yum.

This setup allows for future expansion of classes such as puppetlabs_repo::apt
or ::solaris.

The module is more or less replacing the puppetlabs-prosvc_repo module.

Signed-off-by: Michael Stahnke <stahnma@puppetlabs.com>